scrape: make the initial scrape offset configurable
This change changes the `NoJitter` boolean configuration to a
confiurable `InitialScrapeOffset` option. This will let serverless
setups configure more freely how it will want to scrape workloads.

This was needed because in some deployments in serverless environments
the scraper (colocated with the workload) might not want to wait an
entire scrape interval (since the target could be shortlived) but also
might not want to scrape immediately when the target is not ready.

Making this configurable lets the scrape controller (OpenTelemetry
Contrib's prometheusreceiver) choose that is appropriate for it
depending on its environment.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>